refactor: extract plan instructions + system prompt from streamMessage()
Extract two more cohesive blocks from streamMessage() into a new module:
streamContextBuilder.ts

buildPlanInstructions() handles:
- Plan file reading with legacy migration
- Plan-mode instruction injection
- Plan-file hints in non-plan modes (with Start Here detection)
- Task nesting depth warnings
- Plan→exec handoff transition content determination

buildStreamSystemContext() handles:
- Agent body resolution with inheritance
- Subagent append_prompt resolution
- Available subagent discovery for tool descriptions
- Available skills discovery for tool descriptions
- System message construction
- System message token counting

Also moves discoverAvailableSubagentsForToolContext() helper to the new
module (re-exported from aiService.ts for test compatibility).

Both functions are purely functional — zero this.* dependencies.
15 imports removed from aiService.ts.

Impact: aiService.ts 1574 → 1367 lines (−207)
